+++ golem-worker-service/src/service/api_definition_lookup_impl.rs
@@ -1,0 +1,116 @@
+use crate::worker_bridge_request_executor::WorkerRequestToHttpResponse;
+use async_trait::async_trait;
+use golem_worker_service_base::api_definition::http::HttpApiDefinition;
+use golem_worker_service_base::api_definition::{ApiDefinitionId, ApiVersion};
+use golem_worker_service_base::app_config::WorkerServiceBaseConfig;
+use golem_worker_service_base::auth::{CommonNamespace, EmptyAuthCtx};
+use golem_worker_service_base::http::InputHttpRequest;
+use golem_worker_service_base::repo::api_definition_repo::{
+    ApiDefinitionRepo, InMemoryRegistry, RedisApiRegistry,
+};
+use golem_worker_service_base::service::api_definition::{
+    ApiDefinitionKey, ApiDefinitionService, ApiDefinitionServiceDefault,
+};
+use golem_worker_service_base::service::api_definition_lookup::{
+    ApiDefinitionLookup, ApiDefinitionLookupError,
+};
+use golem_worker_service_base::service::api_definition_validator::ApiDefinitionValidatorNoop;
+use golem_worker_service_base::service::api_definition_validator::ApiDefinitionValidatorService;
+use golem_worker_service_base::service::http::http_api_definition_validator::{
+    HttpApiDefinitionValidator, RouteValidationError,
+};
+use golem_worker_service_base::service::template::{RemoteTemplateService, TemplateServiceNoop};
+use golem_worker_service_base::service::worker::{
+    WorkerRequestMetadata, WorkerServiceDefault, WorkerServiceNoOp,
+};
+use golem_worker_service_base::worker_bridge_execution::WorkerRequestExecutor;
+use http::HeaderMap;
+use poem::Response;
+use std::sync::Arc;
+use tracing::error;
+
+pub struct CustomRequestDefinitionLookupDefault {
+    register_api_definition_repo:
+    Arc<dyn ApiDefinitionRepo<CommonNamespace, HttpApiDefinition> + Sync + Send>,
+}
+
+impl CustomRequestDefinitionLookupDefault {
+    pub fn new(
+        register_api_definition_repo: Arc<
+            dyn ApiDefinitionRepo<CommonNamespace, HttpApiDefinition> + Sync + Send,
+        >,
+    ) -> Self {
+        Self {
+            register_api_definition_repo,
+        }
+    }
+}
+
+#[async_trait]
+impl ApiDefinitionLookup<InputHttpRequest, HttpApiDefinition>
+for CustomRequestDefinitionLookupDefault
+{
+    async fn get(
+        &self,
+        input_http_request: InputHttpRequest,
+    ) -> Result<HttpApiDefinition, ApiDefinitionLookupError> {
+        let api_definition_id = match get_header_value(
+            &input_http_request.headers,
+            "x-golem-api-definition-id", // TODO; This will be removed, and will depend on domain
+        ) {
+            Ok(api_definition_id) => Ok(ApiDefinitionId(api_definition_id.to_string())),
+            Err(err) => Err(ApiDefinitionLookupError(format!(
+                "{} not found in the request headers. Error: {}",
+                "x-golem-api-definition-id", err
+            ))),
+        }?;
+
+        // This will be removed and will be depending on the latest version
+        let version = match get_header_value(
+            &input_http_request.headers,
+            "x-golem-api-definition-version",
+        ) {
+            Ok(version) => Ok(ApiVersion(version)),
+            Err(err) => Err(ApiDefinitionLookupError(format!(
+                "{} not found in the request headers. Error: {}",
+                "x-golem-api-definition-version", err
+            ))),
+        }?;
+
+        let api_key = ApiDefinitionKey {
+            namespace: CommonNamespace::default(),
+            id: api_definition_id.clone(),
+            version: version.clone(),
+        };
+
+        let value = self
+            .register_api_definition_repo
+            .get(&api_key)
+            .await
+            .map_err(|err| {
+                error!("Error getting api definition from the repo: {}", err);
+                ApiDefinitionLookupError(format!(
+                    "Error getting api definition from the repo: {}",
+                    err
+                ))
+            })?;
+
+        value.ok_or(ApiDefinitionLookupError(format!(
+            "Api definition with id: {} and version: {} not found",
+            &api_definition_id, &version
+        )))
+    }
+}
+
+fn get_header_value(headers: &HeaderMap, header_name: &str) -> Result<String, String> {
+    let header_value = headers
+        .iter()
+        .find(|(key, _)| key.as_str().to_lowercase() == header_name)
+        .map(|(_, value)| value)
+        .ok_or(format!("Missing {} header", header_name))?;
+
+    header_value
+        .to_str()
+        .map(|x| x.to_string())
+        .map_err(|e| format!("Invalid value for the header {} error: {}", header_name, e))
+}
